refactor: Line 내 구간 추가 로직 Sections 로 이동
